Implement user_in_region function

+++ c_src/geopoly.h
@@ -23,8 +23,8 @@ typedef struct {
 } GP_Point;
 
 typedef struct {
-  GP_Point *data;
-  size_t size;
+    GP_Point *data;
+    unsigned size;
 } GP_Polygon;
 
 ///
